Basic config
In this first step I inherit a lot of the old Vim configuration from
.vimrc and .vim/ folder. To make the plugins for coding style etc. work
I have a symliink ./after/plugin -> ~/.vim/plugin. But this is a
workaround and should be replaced by LSP and treesitter.

Also the Lazy configuration shall be moved out of the single init.lua
config into something more modular.