Change in lspconfig for clangd
With this patch the filtype is no in the correct position for the setup
and the file compile_commands.json is searched inside the project file
(ChatGPT pointed this out). The compile_commands.json was generated with
bear(1) and needed a little bit of manual patching because it got a
unknown compile parameter.